feat: Add Collection and addBook module
